Security fix for SSH URL components
When calling SSH URLs, Blink will use the URL provided by the system. The URL
percentage encoded components at the Host or user may contain special control
characters. Although those characters may not be allowed on host or user, they
can would make it down and decoded by the shell.

A link may hide this behavior within markdown or even a QR code, so the user
will only see what will run after the fact.

We looked into using libssh parsing uri functions, as this was already a bug at
the ProxyCommand parsing there. Those are not exposed, and are also planned on a
different tokenization and even existing usernames, etc...

This fixes the issue by rejecting sequences with control characters before
sending them to the terminal. The goal is to create valid shell tokens for the
user and host, and then let libssh parse those down the line when calling the
command. We are supporting proper SSH URI separators (;) to connect additional
parameters in the future..

Thanks to David Leadbeater!